Users can only delete products from their own orders

+++ website/views/order_detail_view.py
@@ -4,7 +4,7 @@ from website.models.order_model import Order
 from website.models.product_order_model import ProductOrder
 from website.models.product_model import Product
 from django.contrib.auth.models import User
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseNotFound
 
 
 def order_detail(request, order_id):
@@ -33,24 +33,26 @@ def order_detail(request, order_id):
 	return render(request, template_name, {'order': order, "orderproducts":product_list, "total":total})
 
 def delete_product_from_order(request, product_id, order_id):
-
-	ProductOrder.objects.filter(product_id=product_id, order_id=order_id).delete()
-	# return render(request, 'success.html', {"deleted_object":"Product"})
-	template_name = 'order_detail.html'
 	order = Order.objects.get(pk=order_id)
+	if request.user == order.user:
 
-	# Get seller object
-	line_items = ProductOrder.objects.filter(order=order_id).values_list('product_id').distinct()
-	product_list = list()
-	total = int()
-	for x in line_items:
-		product = Product.objects.filter(pk=x[0])
-		product_count = ProductOrder.objects.filter(product_id=x[0], order=order_id).count()
-		subtotal = product[0].price * product_count
-		total += subtotal
-		product_list.append((product, product_count, subtotal))
-	return render(request, template_name, {'order': order, "orderproducts":product_list, "total":total})
+		ProductOrder.objects.filter(product_id=product_id, order_id=order_id).delete()
+		# return render(request, 'success.html', {"deleted_object":"Product"})
+		template_name = 'order_detail.html'
 
 
-
+		# Get seller object
+		line_items = ProductOrder.objects.filter(order=order_id).values_list('product_id').distinct()
+		product_list = list()
+		total = int()
+		for x in line_items:
+			product = Product.objects.filter(pk=x[0])
+			product_count = ProductOrder.objects.filter(product_id=x[0], order=order_id).count()
+			subtotal = product[0].price * product_count
+			total += subtotal
+			product_list.append((product, product_count, subtotal))
+		return render(request, template_name, {'order': order, "orderproducts":product_list, "total":total})
+	else:
+		# return render(request, 'index.html', {})
+		return HttpResponseNotFound('<h1>Page not found</h1>')
